move data_augmentation to load dataset file
